backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
sqlalchemy
alembic
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
python-dotenv
thefuzz
google-generativeai
pytest
pytest-asyncio
twilio
